Problem: Inconsistent size_t/int usage. size_t and int don't match in 64 bit build and produce warnings. Solution: types corrected.

+++ src/mechanism_base.hpp
@@ -44,7 +44,7 @@ class mechanism_base_t : public mechanism_t
 
     int check_basic_command_structure (msg_t *msg_);
 
-    void handle_error_reason (const char *error_reason, int error_reason_len);
+    void handle_error_reason (const char *error_reason, size_t error_reason_len);
 };
 }
